+++ src/Elders.Cronus.AtomicAction.Redis/RevisionStore/RedisRevisionStore.cs
@@ -1,0 +1,115 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using Elders.Cronus.DomainModeling;
+using Elders.Cronus.Userfull;
+using StackExchange.Redis;
+
+namespace Elders.Cronus.AtomicAction.Redis.RevisionStore
+{
+    public class RedisRevisionStore : IRevisionStore
+    {
+        private ConnectionMultiplexer connection;
+
+        public RedisRevisionStore(IEnumerable<IPEndPoint> redisEndpoints)
+        {
+            if (ReferenceEquals(null, redisEndpoints)) throw new ArgumentNullException(nameof(redisEndpoints));
+
+            var options = new ConfigurationOptions();
+            options.AbortOnConnectFail = false;
+
+            foreach (var endpoint in redisEndpoints)
+            {
+                options.EndPoints.Add(endpoint);
+            }
+
+            connection = ConnectionMultiplexer.Connect(options);
+        }
+
+        public Result<bool> SaveRevision(IAggregateRootId aggregateRootId, int revision)
+        {
+            return SaveRevision(aggregateRootId, revision, null);
+        }
+
+        public Result<bool> SaveRevision(IAggregateRootId aggregateRootId, int revision, TimeSpan? expiry)
+        {
+            if (ReferenceEquals(null, aggregateRootId)) throw new ArgumentNullException(nameof(aggregateRootId));
+
+            if (connection.IsConnected == false)
+                return Result.Error($"Unreachable endpoint '{connection.ClientName}'.");
+
+            var revisionKey = CreateRedisRevisionKey(aggregateRootId);
+
+            try
+            {
+                var result = connection.GetDatabase().StringSet(revisionKey, string.Join(",", revision, DateTime.UtcNow), expiry);
+
+                return new Result<bool>(result);
+            }
+            catch (Exception ex)
+            {
+                return Result.Error(ex);
+            }
+        }
+
+        public Result<int> GetRevision(IAggregateRootId aggregateRootId)
+        {
+            if (ReferenceEquals(null, aggregateRootId)) throw new ArgumentNullException(nameof(aggregateRootId));
+
+            if (connection.IsConnected == false)
+                return new Result<int>().WithError($"Unreachable endpoint '{connection.ClientName}'.");
+
+            var revisionKey = CreateRedisRevisionKey(aggregateRootId);
+
+            try
+            {
+                var value = connection.GetDatabase().StringGet(revisionKey);
+                var revisionValue = ((string)value).Split(',').First();
+
+                return new Result<int>(int.Parse(revisionValue));
+            }
+            catch (Exception ex)
+            {
+                return new Result<int>().WithError(ex);
+            }
+        }
+
+        public Result<bool> HasRevision(IAggregateRootId aggregateRootId)
+        {
+            if (ReferenceEquals(null, aggregateRootId)) throw new ArgumentNullException(nameof(aggregateRootId));
+
+            if (connection.IsConnected == false)
+                return Result.Error($"Unreachable endpoint '{connection.ClientName}'.");
+
+            var revisionKey = CreateRedisRevisionKey(aggregateRootId);
+
+            try
+            {
+                var result = connection.GetDatabase().KeyExists(revisionKey);
+
+                return new Result<bool>(result);
+            }
+            catch (Exception ex)
+            {
+                return Result.Error(ex);
+            }
+        }
+
+        public void Dispose()
+        {
+            if (connection != null)
+            {
+                connection.Dispose();
+                connection = null;
+            }
+        }
+
+        private string CreateRedisRevisionKey(IAggregateRootId aggregateRootId)
+        {
+            var stringRawId = Convert.ToBase64String(aggregateRootId.RawId);
+
+            return string.Concat("revision-", stringRawId);
+        }
+    }
+}