+++ src/Elders.Cronus.AtomicAction.Redis/RELEASE_NOTES.md
@@ -1,2 +1,5 @@
-#### 0.1.0-beta0001 - 09.11.2015
-* Initial release
+#### 0.1.0 - 07.01.2016
+* Updated RedLock reference
+* Added Example project
+* Redis cluster support
+* Fluent configuration for RedisAggregateRootAtomicAction